+++ requirement.txt
@@ -1,0 +1,3 @@
+npm install --save react-modal
+npm install --save react-icons
+npm install react-router-dom
