+++ src/Pages/Cart/Cart.jsx
@@ -1,0 +1,10 @@
+import "./Cart.css"
+function Cart() {
+    return(
+        <div>
+            I am a cart
+        </div>
+    )
+}
+
+export default Cart
