+++ src/Pages/Store/Store.jsx
@@ -1,13 +1,14 @@
 import "./Store.css"
-import Header from "../components/Header/Header"
-import ItemViewer from "../components/ItemViewer/ItemViewer"
+import ItemViewer from "../../components/ItemViewer/ItemViewer"
 function Store() {   
       return (
         <div className="store">
           <div className="store-content">
           <ItemViewer/>
           </div>
-        </div>   
+        </div>
+        
+        
       )
 }
