+++ src/components/CartHeader/CartHeader.css
@@ -3,10 +3,12 @@
     width: fit-content;
     margin: 10px 5px;
     border-radius: 5px;
-    border:solid;
+    border:solid black;
     padding: 5px; 
     gap: 8px;
     position: relative;
+    text-decoration: none;
+    color: black;
 }
 
 .cart-header:hover {
@@ -37,6 +39,8 @@
 }
 
 .cart-block-text {
+    text-decoration: none;
+    color: black;
     font-family: "SF Pro Display", 'Arial', sans-serif;
     font-weight: bold;
     font-size: 0.8rem;
