+++ src/components/CartHeader/CartHeader.jsx
@@ -2,14 +2,14 @@ import "./CartHeader.css"
 function CartHeader() {
     const source_cart_logo = "/cart-logo.png"
     return(
-        <div className="cart-header">
+        <a className="cart-header" href="/cart">
             <img className={"cart-logo"} src={source_cart_logo} alt="Cart"></img>
             <span className="cart-count-icon">0</span>
             <div className="cart-block-text">
                 <span>Giỏ</span> <br/>
                 <span>hàng</span>
             </div>
-        </div>
+        </a>
     );
 };
