+++ src/components/LoginBoard/login.css
@@ -1,0 +1,179 @@
+a {
+  font-weight: 500;
+  color: #646cff;
+  text-decoration: inherit;
+}
+a:hover {
+  color: #800080;
+}
+
+h1 {
+  font-size: 3.2em;
+  line-height: 1.1;
+  text-align: center;
+}
+
+.login-page-container button {
+  background-color: #535bf2;
+  color: #ffffff;
+  border: none;
+  border-radius: 8px;
+  font-size: 1em;
+  font-weight: 500;
+  font-family: inherit;
+  cursor: pointer;
+  transition: background-color 0.25s;
+  justify-content: center;
+  display: flex;
+  align-items: center;
+  padding: 10px;
+  width: 100%;
+}
+.login-page-container button:hover {
+  border-color: #646cff;
+}
+.login-page-container button:focus,
+.login-page-container button:focus-visible {
+  outline: 4px auto -webkit-focus-ring-color;
+}
+
+@media (prefers-color-scheme: light) {
+  :root {
+    color: #213547;
+    background-color: #ffffff;
+  }
+  a:hover {
+    color: #747bff;
+  }
+  button {
+    background-color: #f9f9f9;
+  }
+}
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 80%; /* Make the container take the full width */
+  max-width: 1200px; /* Optional: Set a max width for better readability */
+  margin: 0 auto; /* Center the container horizontally */
+  padding: 20px; /* Optional: Add some padding */
+}
+
+.inputcontainer {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  max-width: 1200px;
+  margin-bottom: 1em;
+}
+
+.inputcontainer input {
+  flex: 1; /* Allow the input to take up the remaining space */
+  padding-left: 2.5em; /* Adjust padding if needed */
+  border: 1px solid #646cff;
+  border-radius: 8px;
+  padding: 0.6em 1.2em;
+  font-size: 1em;
+  font-weight: 500;
+  font-family: inherit;
+  background-color: white;
+  color: #646cff;
+  transition: border-color 0.25s;
+}
+
+.inputform {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-top: 2em;
+  width: 100%;
+}
+
+.inputcontainer input:focus {
+  border-color: #535bf2;
+}
+
+.loginbutton, .signup {
+  margin-top: 1em;
+}
+
+.registerredirect, .loginredirect {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px;
+  width: 100%;
+  max-width: 400px; /* Optional: Set a max width for better readability */
+}
+
+.registerredirect button, .loginredirect button {
+  margin-left: auto; /* Push the button to the far right */
+}
+
+.forgotpassword {
+  margin-top: 10px; /* Add some margin if needed */
+}
+
+.forgotpassword:hover {
+  color: #800080;
+}
+.signup {
+  margin-top: 1em;
+}
+.loginredirect {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding-top: 10px;
+  width: 100%;
+  max-width: 1200px; /* Optional: Set a max width for better readability */
+}
+
+.loginredirect p {
+  margin: 0; /* Remove default margin */
+}
+
+.loginredirect button {
+  margin-left: 20px; /* Push the button to the far right */
+}
+.inputcontainer img {
+  margin-right: 0.5em; /* Add some space between the icon and the input */
+  width: 2.5em; /* Adjust the size of the icon */
+  height: auto;
+}
+/* Add these styles to your login.css or a new CSS file */
+.login-modal {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  right: auto;
+  bottom: auto;
+  margin-right: -50%;
+  transform: translate(-50%, -50%);
+  padding: 20px;
+  border-radius: 10px;
+  outline: none;
+  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
+  line-height: 1.5;
+  font-weight: 400;
+  color-scheme: light dark;
+  color: rgba(0, 0, 0, 0.87);
+  background-color: white;
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  width: 100%; /* Adjust the width as needed */
+  max-width: 400px; /* Set a max width for the modal */
+}
+
+.login-modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.75);
+}
