+++ src/components/LoginBoard/loginapp.jsx
@@ -1,0 +1,23 @@
+import React, { useState } from 'react';
+import Login from './login.jsx';
+import Register from './register.jsx';
+
+function loginapp() {
+    const [isRegister, setIsRegister] = useState(false);
+
+    const handleRegisterClick = () => {
+        setIsRegister(true);
+    };
+
+    const handleLoginClick = () => {
+        setIsRegister(false);
+    };
+
+    return (
+        <div>
+            {isRegister ? <Register onLoginClick={handleLoginClick} /> : <Login onRegisterClick={handleRegisterClick} />}
+        </div>
+    );
+}
+
+export default loginapp;
