+++ src/components/SearchBar/SearchBar.css
@@ -13,7 +13,9 @@
     height: 50%;
     padding-right: 10px;
     opacity: 50%;
+    cursor: pointer;
 }
+
 
 .search-input {
     margin-right: 10px;
